delete ls_file_list and fix ls dependency resolution
